pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agri-perception-person-detection"
version = "0.1.0"
description = "Safety-critical person detection pipeline for agricultural autonomy"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Project Contributor" }
]
dependencies = [
  "ultralytics>=8.2.0",
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "numpy>=1.24.0",
  "opencv-python>=4.8.0",
  "albumentations>=1.4.0",
  "pycocotools>=2.0.7",
  "matplotlib>=3.7.0",
  "pillow>=10.0.0",
  "pyyaml>=6.0.0",
  "pandas>=2.0.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
